修复内存厂商信息显示错误问题

+++ GrantleySocketPkg/Library/MemoryQpiInit/Platform/Hooks/CommonHooks.c
@@ -146,7 +146,7 @@ PlatformReadSmb (
   PSYSHOST         host,
   UINT8            socket,
   struct smbDevice dev,
-  UINT8            byteOffset,
+  UINT16            byteOffset,
   UINT8            *data
   )
 /*++
